Perf: Optimizing AtomicIntHashCounter and AccurateRecorder
