+++ MyPerf4J-Benchmark/src/main/java/cn/myperf4j/bench/util/concurrent/AtomicIntHashCounterBench.java
@@ -1,8 +1,8 @@
 package cn.myperf4j.bench.util.concurrent;
 
 import cn.myperf4j.base.util.concurrent.AtomicIntArray;
+import cn.myperf4j.base.util.concurrent.IntHashCounter;
 import cn.myperf4j.base.util.concurrent.AtomicIntHashCounter;
-import cn.myperf4j.base.util.concurrent.ScalableAtomicIntHashCounter;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Level;
@@ -25,13 +25,48 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by LinShunkang on 2022/03/20
+ * <p>
+ * # JMH version: 1.34
+ * # VM version: JDK 17.0.4, Java HotSpot(TM) 64-Bit Server VM, 17.0.4+11-LTS-179
+ * # VM invoker: /Library/Java/JavaVirtualMachines/jdk-17.0.4.jdk/Contents/Home/bin/java
+ * # VM options: -ea -Xmx8G -Xms8G -Xmn4G -javaagent:/Applications/IntelliJ IDEA.app/Contents/lib/idea_rt.jar=52807:/Applications/IntelliJ IDEA.app/Contents/bin -Dfile.encoding=UTF-8
+ * # Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
+ * # Warmup: 1 iterations, 10 s each
+ * # Measurement: 5 iterations, 10 s each
+ * # Timeout: 10 min per iteration
+ * # Benchmark mode: Throughput, ops/time
+ * # Parameters: (mapSize = 1048576)
+ * <p>
+ * # Threads: 1 thread
+ * Benchmark                                         (mapSize)   Mode  Cnt    Score    Error   Units
+ * AtomicIntHashCounterBench.intArray                  1048576  thrpt    5  241.139 ±  2.975  ops/us
+ * AtomicIntHashCounterBench.jdkIntegerMap             1048576  thrpt    5   13.969 ±  3.741  ops/us
+ * AtomicIntHashCounterBench.scalableIntHashCounter    1048576  thrpt    5  184.314 ± 32.741  ops/us
+ * <p>
+ * # Threads: 2 thread
+ * Benchmark                                         (mapSize)   Mode  Cnt    Score    Error   Units
+ * AtomicIntHashCounterBench.intArray                  1048576  thrpt    5  375.245 ± 23.964  ops/us
+ * AtomicIntHashCounterBench.jdkIntegerMap             1048576  thrpt    5   33.437 ±  7.945  ops/us
+ * AtomicIntHashCounterBench.scalableIntHashCounter    1048576  thrpt    5  327.925 ± 44.565  ops/us
+ * <p>
+ * # Threads: 4 thread
+ * Benchmark                                         (mapSize)   Mode  Cnt    Score    Error   Units
+ * AtomicIntHashCounterBench.intArray                  1048576  thrpt    5  532.269 ± 21.093  ops/us
+ * AtomicIntHashCounterBench.jdkIntegerMap             1048576  thrpt    5   67.699 ±  2.826  ops/us
+ * AtomicIntHashCounterBench.scalableIntHashCounter    1048576  thrpt    5  503.602 ± 61.535  ops/us
+ * <p>
+ * # Threads: 8 thread
+ * Benchmark                                         (mapSize)   Mode  Cnt    Score   Error   Units
+ * AtomicIntHashCounterBench.intArray                  1048576  thrpt    5   86.101 ± 8.562  ops/us
+ * AtomicIntHashCounterBench.jdkIntegerMap             1048576  thrpt    5  133.705 ± 3.813  ops/us
+ * AtomicIntHashCounterBench.scalableIntHashCounter    1048576  thrpt    5   86.922 ± 6.591  ops/us
  */
 @BenchmarkMode({Mode.Throughput})
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
 public class AtomicIntHashCounterBench {
 
-    private AtomicIntHashCounter scalableIntHashCounter;
+    private IntHashCounter intHashCounter;
 
     private AtomicIntArray intArray;
 
@@ -42,14 +77,14 @@ public class AtomicIntHashCounterBench {
 
     @Setup(Level.Iteration)
     public void setup() {
-        scalableIntHashCounter = new ScalableAtomicIntHashCounter(128);
+        intHashCounter = new AtomicIntHashCounter(128);
         intArray = new AtomicIntArray(mapSize + 1);
         jdkIntegerMap = new ConcurrentHashMap<>(128);
 
         final ThreadLocalRandom random = ThreadLocalRandom.current();
         for (int i = 1; i < mapSize; i++) {
             final int key = random.nextInt(0, mapSize);
-            scalableIntHashCounter.incrementAndGet(key);
+            intHashCounter.incrementAndGet(key);
             intArray.incrementAndGet(key);
             increase(jdkIntegerMap, key);
         }
@@ -61,7 +96,7 @@ public class AtomicIntHashCounterBench {
 
     @Benchmark
     public int scalableIntHashCounter(ThreadState state) {
-        return scalableIntHashCounter.incrementAndGet(randomKey(state));
+        return intHashCounter.incrementAndGet(randomKey(state));
     }
 
     @Benchmark
